+++ source/SoundTouch/InterpolateShannon.h
@@ -61,7 +61,7 @@ protected:
                         const SAMPLETYPE *src, 
                         int &srcSamples);
 
-    float fract;
+    double fract;
 
 public:
     InterpolateShannon();
